made it compile with different compiler versions under lenny (gtk 2.12); fixed importing of preset bank with different version (destroyed main settings); convert jconv.wet_dry to new range; no error message when no ir-file set; fixed switches; fixed double loading of presets; better numerical stability for gainline calculation; added generated faust tonestack files; fixed possible crash in IRWindow; removed some debug output; preset format version incremented (1.2); really exclude debian dir when --no-pkg set; fixed manual entry of delay/offset in IRWindow; guitarix version 0.11.0

+++ trunk/libgxw/gxw/GxEQSlider.cpp
@@ -65,7 +65,7 @@ static void eq_slider_expose(
 	GtkWidget *widget, GdkRectangle *rect, gdouble sliderstate, GdkPixbuf *image)
 {
 	gdk_draw_pixbuf(GDK_DRAWABLE(widget->window), widget->style->fg_gc[0],
-	                image, 0, sliderstate, rect->x, rect->y,
+	                image, 0, (gint)sliderstate, rect->x, rect->y,
 	                rect->width, rect->height, GDK_RGB_DITHER_NORMAL, 0, 0);
 }
